+++ backend/canisters/community/impl/src/updates/c2c_install_bot.rs
@@ -1,25 +1,24 @@
+use crate::guards::caller_is_local_user_index;
 use crate::{activity_notifications::handle_activity_notification, mutate_state, run_regular_jobs, RuntimeState};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use group_canister::add_bot::{Response::*, *};
+use types::c2c_install_bot::{Response::*, *};
 use types::BotGroupConfig;
 
-#[update(msgpack = true)]
+#[update(guard = "caller_is_local_user_index", msgpack = true)]
 #[trace]
-fn add_bot(args: Args) -> Response {
+fn c2c_install_bot(args: Args) -> Response {
     run_regular_jobs();
 
-    mutate_state(|state| add_bot_impl(args, state))
+    mutate_state(|state| c2c_install_bot_impl(args, state))
 }
 
-fn add_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
+fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
     if state.data.is_frozen() {
-        return ChatFrozen;
+        return Frozen;
     }
 
-    let caller = state.env.caller();
-
-    let Some(member) = state.data.get_member(caller) else {
+    let Some(member) = state.data.members.get_by_user_id(&args.caller) else {
         return NotAuthorized;
     };
 
@@ -27,8 +26,8 @@ fn add_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
         return NotAuthorized;
     }
 
-    if !state.data.add_bot(
-        member.user_id(),
+    if !state.data.install_bot(
+        member.user_id,
         args.bot_id,
         BotGroupConfig {
             permissions: args.granted_permissions,
@@ -37,8 +36,6 @@ fn add_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
     ) {
         return AlreadyAdded;
     }
-
-    // TODO: Notify UserIndex
 
     handle_activity_notification(state);
     Success