+++ backend/libraries/types/src/c2c_install_bot.rs
@@ -1,17 +1,18 @@
+use crate::{SlashCommandPermissions, UserId};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use ts_export::ts_export;
-use types::UserId;
 
-#[ts_export(group, remove_bot)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub bot_id: UserId,
+    pub caller: UserId,
+    pub granted_permissions: SlashCommandPermissions,
 }
 
-#[ts_export(group, remove_bot)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
+    Frozen,
     NotAuthorized,
+    AlreadyAdded,
 }